Optimizations - add HWord - add functions that replace LINQ queries against lists - allow fixed lists to start from an index other than 0 - remove unused code

+++ Source/Roton.Test/Roton/Integration/Elements/PlayerTests.cs
@@ -247,7 +247,7 @@ public class PlayerTests(Context context) : ElementTestFixture(context)
             "bomb should be present after activation");
         Message.Should().BeEquivalentTo(Alerts.BombMessage.Text,
             "correct message should be displayed");
-        actor.P1.Should().Be((byte)(Engine.Facts.BombCountdownStart - 1),
+        ((int)actor.P1).Should().Be((byte)(Engine.Facts.BombCountdownStart - 1),
             "bomb should have the maximum timer set");
     }
